+++ src/libecalc/domain/infrastructure/energy_components/legacy_consumer/consumer_function_mapper/__init__.py
@@ -6,10 +6,10 @@ from libecalc.common.logger import logger
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_function import ConsumerFunction
 from libecalc.domain.process.compressor.dto.base import CompressorConsumerFunction
 from libecalc.domain.process.dto import (
+    DirectConsumerFunction,
     TabulatedConsumerFunction,
 )
 from libecalc.domain.process.dto.consumer_system import CompressorSystemConsumerFunction
-from libecalc.domain.process.dto.direct import DirectConsumerFunction
 from libecalc.domain.process.pump.pump_consumer_function import PumpConsumerFunction
 
 from .compressor_consumer_function import create_compressor_consumer_function
