+++ src/libecalc/domain/process/compressor/core/train/variable_speed_compressor_train_common_shaft_multiple_streams_and_pressures.py
@@ -28,8 +28,8 @@ from libecalc.domain.process.compressor.core.train.utils.numeric_methods import 
 from libecalc.domain.process.compressor.core.train.utils.variable_speed_compressor_train_common_shaft import (
     get_single_speed_equivalent,
 )
-from libecalc.domain.process.compressor.dto.stage import CompressorStage
-from libecalc.domain.process.compressor.dto.train import (
+from libecalc.domain.process.compressor.dto import (
+    CompressorStage,
     SingleSpeedCompressorTrain,
     VariableSpeedCompressorTrainMultipleStreamsAndPressures,
 )
