+++ src/libecalc/domain/process/compressor/core/utils.py
@@ -10,8 +10,7 @@ from libecalc.domain.process.compressor.core.train.stage import (
     CompressorTrainStage,
     UndefinedCompressorStage,
 )
-from libecalc.domain.process.compressor.dto.chart import CompressorChart
-from libecalc.domain.process.compressor.dto.stage import CompressorStage
+from libecalc.domain.process.compressor.dto import CompressorChart, CompressorStage
 from libecalc.domain.process.dto import (
     GenericChartFromDesignPoint,
     GenericChartFromInput,
